# services/order-service/swagger.yaml
openapi: 3.0.3
info:
  title: SmartRetailX Order Service
  description: Create and manage orders; integrates with catalogue and payment via REST
  version: 1.0.0
servers:
  - url: /api/v1
    description: Versioned API
tags:
  - name: Orders
paths:
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: userId
          schema:
            type: string
      responses:
        '200':
          description: Order list
    post:
      tags: [Orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
        '400':
          description: Validation or catalogue error
        '502':
          description: Downstream service failure
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order
        '404':
          description: Not found
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status (publishes event)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Updated order
        '400':
          description: Invalid status transition
components:
  schemas:
    CreateOrderRequest:
      type: object
      required: [userId, items, paymentMethod]
      properties:
        userId:
          type: string
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
              quantity:
                type: integer
                minimum: 1
        paymentMethod:
          type: string
          enum: [card, wallet, cash_on_delivery]
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]